Remove further unwanted linker flags from perl_embed_ldflags
Remove the contents of $Config{ldflags} from ExtUtils::Embed's ldopts,
like we already do with $Config{ccdlflags}.  Those flags the choices
of those who built the Perl installation, which are not necessarily
appropriate for building PostgreSQL.  What we really want from ldopts
are the options identifying the location and name of the libperl
library, but unfortunately it doesn't appear possible to get that
separately from the other stuff.

The motivation for this was to strip -mmacosx-version-min options.  We
already did something similar for the -arch option.  Both of those are
now covered by this more general approach.